[ERMP] update styles. Move errorBoundary into separate componen
