+++ src/components/Footer/index.jsx
@@ -1,14 +1,17 @@
 import React from 'react';
 import Wrapper from '../Wrapper';
+import ErrorBoundary from '../ErrorBoundary';
 import './styles.css';
 
 export default () => (
   <footer>
-    <Wrapper>
-      <p className="app-title">
-        <span className="app-title__name">netflix</span>
-        roulette
-      </p>
-    </Wrapper>
+    <ErrorBoundary>
+      <Wrapper>
+        <p className="app-title">
+          <span className="app-title__name">netflix</span>
+          roulette
+        </p>
+      </Wrapper>
+    </ErrorBoundary>
   </footer>
 );
