+++ src/components/Header/Brand/styles.css
@@ -1,5 +1,5 @@
 .brand {
-  height: 20%;
+  height: 100%;
   display: flex;
   align-items: center;
 }
