+++ src/components/Wrapper/index.jsx
@@ -1,30 +1,8 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import './styles.css';
 
-class Wrapper extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { hasError: false };
-  }
-
-  static getDerivedStateFromError() {
-    return { hasError: true };
-  }
-
-  render() {
-    const { hasError } = this.state;
-    const { children } = this.props;
-    if (hasError) {
-      return (
-        <div className="error">
-          Something went wrong. Please try again later
-        </div>
-      );
-    }
-    return <div className="wrapper">{children}</div>;
-  }
-}
+const Wrapper = ({ children }) => <div className="wrapper">{children}</div>;
 
 Wrapper.propTypes = {
   children: PropTypes.oneOfType([
