+++ src/components/index.js
@@ -1,5 +1,6 @@
 import Footer from './Footer';
 import Header from './Header';
 import Main from './Main';
+import TopSection from './TopSection';
 
-export { Footer, Header, Main };
+export { Footer, Header, Main, TopSection };
